fix: move back `index.js` to root
