+++ eslint.config.js
@@ -9,7 +9,7 @@
  * OF ANY KIND, either express or implied. See the License for the specific language
  * governing permissions and limitations under the License.
  */
-import { common } from './src/index.js';
+import { common } from './index.js';
 
 export default [
   common,
