+++ index.js
@@ -11,16 +11,16 @@
  */
 import globals from 'globals';
 
-import bestPractices from '../rules/best-practices.js';
-import errors from '../rules/errors.js';
-import node from '../rules/node.js';
-import style from '../rules/style.js';
-import variables from '../rules/variables.js';
-import es6 from '../rules/es6.js';
-import imports from '../rules/imports.js';
-import strict from '../rules/strict.js';
+import bestPractices from './rules/best-practices.js';
+import errors from './rules/errors.js';
+import node from './rules/node.js';
+import style from './rules/style.js';
+import variables from './rules/variables.js';
+import es6 from './rules/es6.js';
+import imports from './rules/imports.js';
+import strict from './rules/strict.js';
 
-import header from '../rules/header.js';
+import header from './rules/header.js';
 
 const common = {
   languageOptions: {
