+++ package.json
@@ -2,7 +2,7 @@
   "name": "@adobe/eslint-config-helix",
   "version": "2.0.9",
   "description": "Helix's ESLint config, based on Airbnb's style guide",
-  "main": "src/index.js",
+  "main": "index.js",
   "type": "module",
   "scripts": {
     "lint": "./node_modules/.bin/eslint .",
